More size_t
git-svn-id: https://svn.blender.org/svnroot/verse/trunk@74 49c17534-1d02-dc11-b283-0030488c597c

+++ vs_connection.c
@@ -124,7 +124,7 @@ void vs_remove_subscriptor(VSSubscriptionList *list)
 		list->session[i] = list->session[--list->session_count];
 }
 
-unsigned int vs_get_subscript_count(VSSubscriptionList *list)
+size_t vs_get_subscript_count(VSSubscriptionList *list)
 {
 	return list->session_count;
 }
